+++ src/reducers/index.js
@@ -1,5 +1,4 @@
 import { combineReducers } from 'redux';
-import { routerReducer } from 'react-router-redux';
 import { reducer as formReducer } from 'redux-form';
 
 import entities from './entities';
@@ -10,7 +9,6 @@ const rootReducer = combineReducers({
   entities,
   githubProjects,
   auth,
-  routing: routerReducer,
   form: formReducer
 });